Initial commit
This project called Quiz app was sponsored by google and support by andela company. The aim of the project is to design a quiz app that can take user input as answer and give out the result when the user clicks the submit button.Here is the link  for the app using google drive. https://drive.google.com/open?id=11YQ5yd1OP3kriDHiKs-tMIldx5ZIRQab